+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allow underscores (`_`) in HTTP header names.
 - Fixed <kbd>C-u C-c C-r C-r</kbd> and <kbd>C-u C-c C-r C-s</kbd> not displaying the response buffer when a response is received.
 - Fixed Verb not reading parent headings when buffer has been narrowed (e.g. with `org-narrow-to-subtree`).
+- When using Babel `verb` source blocks, arguments specified via the `:var` keyword can now be read using `verb-var`.
 
 ## **2.13.1** - 2021-01-12 (MELPA Stable)
 - Heading properties (e.g. `Verb-Store`, `Verb-Map-Request`) can now be inherited from parent headings by setting `org-use-property-inheritance` to `t`.